+++ R/zzz.R
@@ -11,4 +11,9 @@
 
   packageStartupMessage(paste0(msg, citation))
 
+  options(ChIPseeker.downstreamDistance = 300)
+  options(ChIPseeker.ignore_1st_exon = FALSE)
+  options(ChIPseeker.ignore_1st_intron = FALSE)
+  options(ChIPseeker.ignore_downstream = FALSE)
+  options(ChIPseeker.ignore_promoter_subcategory= FALSE)
 }