+++ NEWS.md
@@ -1,5 +1,8 @@
 # ChIPseeker 1.15.3
 
++ options to ignore '1st exon', '1st intron', 'downstream' and promoter
+  subcategory when summarizing result and visualization (2018-01-09, Tue)
+      - <https://support.bioconductor.org/p/104676/#104689>
 + throw msg of 'file not found and skip' when requested url is not available
   when downloading BED file from GEO (2017-12-28, Thu)
       - <https://support.bioconductor.org/p/104491/#104507>